+++ capa/features/extractors/binja/insn.py
@@ -27,10 +27,28 @@ from binaryninja import (
 )
 
 import capa.features.extractors.helpers
-from capa.features.insn import API, MAX_STRUCTURE_SIZE, Number, Offset, Mnemonic, OperandNumber, OperandOffset
-from capa.features.common import MAX_BYTES_FEATURE_SIZE, Bytes, String, Feature, Characteristic
+from capa.features.insn import (
+    API,
+    MAX_STRUCTURE_SIZE,
+    Number,
+    Offset,
+    Mnemonic,
+    OperandNumber,
+    OperandOffset,
+)
+from capa.features.common import (
+    MAX_BYTES_FEATURE_SIZE,
+    Bytes,
+    String,
+    Feature,
+    Characteristic,
+)
 from capa.features.address import Address, AbsoluteVirtualAddress
-from capa.features.extractors.binja.helpers import DisassemblyInstruction, visit_llil_exprs, get_llil_instr_at_addr
+from capa.features.extractors.binja.helpers import (
+    DisassemblyInstruction,
+    visit_llil_exprs,
+    get_llil_instr_at_addr,
+)
 from capa.features.extractors.base_extractor import BBHandle, InsnHandle, FunctionHandle
 
 # security cookie checks may perform non-zeroing XORs, these are expected within a certain
@@ -63,7 +81,9 @@ def is_stub_function(bv: BinaryView, addr: int) -> Optional[int]:
     return llil.dest.constant
 
 
-def extract_insn_api_features(fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle) -> Iterator[tuple[Feature, Address]]:
+def extract_insn_api_features(
+    fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle
+) -> Iterator[tuple[Feature, Address]]:
     """
     parse instruction API features
 
@@ -112,11 +132,15 @@ def extract_insn_api_features(fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle)
                         elif lib_name.endswith(".so"):
                             lib_name = lib_name[:-3]
 
-                    for name in capa.features.extractors.helpers.generate_symbols(lib_name, sym_name):
+                    for name in capa.features.extractors.helpers.generate_symbols(
+                        lib_name, sym_name
+                    ):
                         yield API(name), ih.address
 
                     if sym_name.startswith("_"):
-                        for name in capa.features.extractors.helpers.generate_symbols(lib_name, sym_name[1:]):
+                        for name in capa.features.extractors.helpers.generate_symbols(
+                            lib_name, sym_name[1:]
+                        ):
                             yield API(name), ih.address
 
 
@@ -132,17 +156,31 @@ def extract_insn_number_features(
 
     results: list[tuple[Any[Number, OperandNumber], Address]] = []
 
-    def llil_checker(il: LowLevelILInstruction, parent: LowLevelILInstruction, index: int) -> bool:
+    def llil_checker(
+        il: LowLevelILInstruction, parent: LowLevelILInstruction, index: int
+    ) -> bool:
         if il.operation == LowLevelILOperation.LLIL_LOAD:
             return False
 
-        if il.operation not in [LowLevelILOperation.LLIL_CONST, LowLevelILOperation.LLIL_CONST_PTR]:
+        if il.operation not in [
+            LowLevelILOperation.LLIL_CONST,
+            LowLevelILOperation.LLIL_CONST_PTR,
+        ]:
             return True
 
         for op in parent.operands:
-            if isinstance(op, ILRegister) and op.name in ["esp", "ebp", "rsp", "rbp", "sp"]:
+            if isinstance(op, ILRegister) and op.name in [
+                "esp",
+                "ebp",
+                "rsp",
+                "rbp",
+                "sp",
+            ]:
                 return False
-            elif isinstance(op, LowLevelILInstruction) and op.operation == LowLevelILOperation.LLIL_REG:
+            elif (
+                isinstance(op, LowLevelILInstruction)
+                and op.operation == LowLevelILOperation.LLIL_REG
+            ):
                 if op.src.name in ["esp", "ebp", "rsp", "rbp", "sp"]:
                     return False
 
@@ -161,7 +199,9 @@ def extract_insn_number_features(
     yield from results
 
 
-def extract_insn_bytes_features(fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle) -> Iterator[tuple[Feature, Address]]:
+def extract_insn_bytes_features(
+    fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle
+) -> Iterator[tuple[Feature, Address]]:
     """
     parse referenced byte sequences
     example:
@@ -173,7 +213,10 @@ def extract_insn_bytes_features(fh: FunctionHandle, bbh: BBHandle, ih: InsnHandl
     candidate_addrs = set()
 
     llil = func.get_llil_at(ih.address)
-    if llil is None or llil.operation in [LowLevelILOperation.LLIL_CALL, LowLevelILOperation.LLIL_CALL_STACK_ADJUST]:
+    if llil is None or llil.operation in [
+        LowLevelILOperation.LLIL_CALL,
+        LowLevelILOperation.LLIL_CALL_STACK_ADJUST,
+    ]:
         return
 
     for ref in bv.get_code_refs_from(ih.address):
@@ -186,8 +229,13 @@ def extract_insn_bytes_features(fh: FunctionHandle, bbh: BBHandle, ih: InsnHandl
         candidate_addrs.add(ref)
 
     # collect candidate address by enumerating all integers, https://github.com/Vector35/binaryninja-api/issues/3966
-    def llil_checker(il: LowLevelILInstruction, parent: LowLevelILInstruction, index: int) -> bool:
-        if il.operation in [LowLevelILOperation.LLIL_CONST, LowLevelILOperation.LLIL_CONST_PTR]:
+    def llil_checker(
+        il: LowLevelILInstruction, parent: LowLevelILInstruction, index: int
+    ) -> bool:
+        if il.operation in [
+            LowLevelILOperation.LLIL_CONST,
+            LowLevelILOperation.LLIL_CONST_PTR,
+        ]:
             value = il.value.value
             if value > 0:
                 candidate_addrs.add(value)
@@ -200,7 +248,9 @@ def extract_insn_bytes_features(fh: FunctionHandle, bbh: BBHandle, ih: InsnHandl
 
     for addr in candidate_addrs:
         extracted_bytes = bv.read(addr, MAX_BYTES_FEATURE_SIZE)
-        if extracted_bytes and not capa.features.extractors.helpers.all_zeros(extracted_bytes):
+        if extracted_bytes and not capa.features.extractors.helpers.all_zeros(
+            extracted_bytes
+        ):
             if bv.get_string_at(addr) is None:
                 # don't extract byte features for obvious strings
                 yield Bytes(extracted_bytes), ih.address
@@ -231,8 +281,13 @@ def extract_insn_string_features(
         candidate_addrs.add(ref)
 
     # collect candidate address by enumerating all integers, https://github.com/Vector35/binaryninja-api/issues/3966
-    def llil_checker(il: LowLevelILInstruction, parent: LowLevelILInstruction, index: int) -> bool:
-        if il.operation in [LowLevelILOperation.LLIL_CONST, LowLevelILOperation.LLIL_CONST_PTR]:
+    def llil_checker(
+        il: LowLevelILInstruction, parent: LowLevelILInstruction, index: int
+    ) -> bool:
+        if il.operation in [
+            LowLevelILOperation.LLIL_CONST,
+            LowLevelILOperation.LLIL_CONST_PTR,
+        ]:
             value = il.value.value
             if value > 0:
                 candidate_addrs.add(value)
@@ -277,13 +332,21 @@ def extract_insn_offset_features(
     results: list[tuple[Any[Offset, OperandOffset], Address]] = []
     address_size = func.view.arch.address_size * 8
 
-    def llil_checker(il: LowLevelILInstruction, parent: LowLevelILInstruction, index: int) -> bool:
+    def llil_checker(
+        il: LowLevelILInstruction, parent: LowLevelILInstruction, index: int
+    ) -> bool:
         #  The most common case, read/write dereference to something like `dword [eax+0x28]`
         if il.operation in [LowLevelILOperation.LLIL_ADD, LowLevelILOperation.LLIL_SUB]:
             left = il.left
             right = il.right
             # Exclude offsets based on stack/franme pointers
-            if left.operation == LowLevelILOperation.LLIL_REG and left.src.name in ["esp", "ebp", "rsp", "rbp", "sp"]:
+            if left.operation == LowLevelILOperation.LLIL_REG and left.src.name in [
+                "esp",
+                "ebp",
+                "rsp",
+                "rbp",
+                "sp",
+            ]:
                 return True
 
             if right.operation != LowLevelILOperation.LLIL_CONST:
@@ -294,13 +357,20 @@ def extract_insn_offset_features(
             # [0, MAX_STRUCTURE_SIZE]. For example,
             # add eax, 0x10,
             # lea ebx, [eax + 1]
-            if parent.operation not in [LowLevelILOperation.LLIL_LOAD, LowLevelILOperation.LLIL_STORE]:
-                if il.operation != LowLevelILOperation.LLIL_ADD or (not 0 < raw_value < MAX_STRUCTURE_SIZE):
+            if parent.operation not in [
+                LowLevelILOperation.LLIL_LOAD,
+                LowLevelILOperation.LLIL_STORE,
+            ]:
+                if il.operation != LowLevelILOperation.LLIL_ADD or (
+                    not 0 < raw_value < MAX_STRUCTURE_SIZE
+                ):
                     return False
 
             if address_size > 0:
                 # BN also encodes the constant value as two's complement, we need to restore its original value
-                value = capa.features.extractors.helpers.twos_complement(raw_value, address_size)
+                value = capa.features.extractors.helpers.twos_complement(
+                    raw_value, address_size
+                )
             else:
                 value = raw_value
 
@@ -309,7 +379,10 @@ def extract_insn_offset_features(
             return False
 
         # An edge case: for code like `push dword [esi]`, we need to generate a feature for offset 0x0
-        elif il.operation in [LowLevelILOperation.LLIL_LOAD, LowLevelILOperation.LLIL_STORE]:
+        elif il.operation in [
+            LowLevelILOperation.LLIL_LOAD,
+            LowLevelILOperation.LLIL_STORE,
+        ]:
             if il.operands[0].operation == LowLevelILOperation.LLIL_REG:
                 results.append((Offset(0), ih.address))
                 results.append((OperandOffset(index, 0), ih.address))
@@ -323,7 +396,9 @@ def extract_insn_offset_features(
     yield from results
 
 
-def is_nzxor_stack_cookie(f: Function, bb: bn.BasicBlock, llil: LowLevelILInstruction) -> bool:
+def is_nzxor_stack_cookie(
+    f: Function, bb: bn.BasicBlock, llil: LowLevelILInstruction
+) -> bool:
     """check if nzxor exists within stack cookie delta"""
     # TODO(xusheng): use LLIL SSA to do more accurate analysis
     # https://github.com/mandiant/capa/issues/1609
@@ -340,11 +415,15 @@ def is_nzxor_stack_cookie(f: Function, bb: bn.BasicBlock, llil: LowLevelILInstru
         return False
 
     # expect security cookie init in first basic block within first bytes (instructions)
-    if len(bb.incoming_edges) == 0 and llil.address < (bb.start + SECURITY_COOKIE_BYTES_DELTA):
+    if len(bb.incoming_edges) == 0 and llil.address < (
+        bb.start + SECURITY_COOKIE_BYTES_DELTA
+    ):
         return True
 
     # ... or within last bytes (instructions) before a return
-    if len(bb.outgoing_edges) == 0 and llil.address > (bb.end - SECURITY_COOKIE_BYTES_DELTA):
+    if len(bb.outgoing_edges) == 0 and llil.address > (
+        bb.end - SECURITY_COOKIE_BYTES_DELTA
+    ):
         return True
 
     return False
@@ -361,7 +440,9 @@ def extract_insn_nzxor_characteristic_features(
 
     results = []
 
-    def llil_checker(il: LowLevelILInstruction, parent: LowLevelILInstruction, index: int) -> bool:
+    def llil_checker(
+        il: LowLevelILInstruction, parent: LowLevelILInstruction, index: int
+    ) -> bool:
         # If the two operands of the xor instruction are the same, the LLIL will be translated to other instructions,
         # e.g., <llil: eax = 0>, (LLIL_SET_REG). So we do not need to check whether the two operands are the same.
         if il.operation == LowLevelILOperation.LLIL_XOR:
@@ -409,7 +490,9 @@ def extract_insn_peb_access_characteristic_features(
 
     results = []
 
-    def llil_checker(il: LowLevelILInstruction, parent: LowLevelILOperation, index: int) -> bool:
+    def llil_checker(
+        il: LowLevelILInstruction, parent: LowLevelILOperation, index: int
+    ) -> bool:
         if il.operation != LowLevelILOperation.LLIL_LOAD:
             return True
 
@@ -449,7 +532,9 @@ def extract_insn_segment_access_features(
 
     results = []
 
-    def llil_checker(il: LowLevelILInstruction, parent: LowLevelILInstruction, index: int) -> bool:
+    def llil_checker(
+        il: LowLevelILInstruction, parent: LowLevelILInstruction, index: int
+    ) -> bool:
         if il.operation == LowLevelILOperation.LLIL_REG:
             reg = il.src.name
             if reg == "fsbase":
@@ -490,7 +575,9 @@ def extract_insn_cross_section_cflow(
             yield Characteristic("cross section flow"), ih.address
 
 
-def extract_function_calls_from(fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle) -> Iterator[tuple[Feature, Address]]:
+def extract_function_calls_from(
+    fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle
+) -> Iterator[tuple[Feature, Address]]:
     """extract functions calls from features
 
     most relevant at the function scope, however, its most efficient to extract at the instruction scope
@@ -514,21 +601,27 @@ def extract_function_calls_from(fh: FunctionHandle, bbh: BBHandle, ih: InsnHandl
             value = dest.value.value
             yield Characteristic("calls from"), AbsoluteVirtualAddress(value)
         elif dest.operation == LowLevelILOperation.LLIL_CONST:
-            yield Characteristic("calls from"), AbsoluteVirtualAddress(dest.value)
+            yield Characteristic("calls from"), AbsoluteVirtualAddress(dest.value.value)
         elif dest.operation == LowLevelILOperation.LLIL_LOAD:
             indirect_src = dest.src
             if indirect_src.operation == LowLevelILOperation.LLIL_CONST_PTR:
                 value = indirect_src.value.value
                 yield Characteristic("calls from"), AbsoluteVirtualAddress(value)
             elif indirect_src.operation == LowLevelILOperation.LLIL_CONST:
-                yield Characteristic("calls from"), AbsoluteVirtualAddress(indirect_src.value)
+                yield (
+                    Characteristic("calls from"),
+                    AbsoluteVirtualAddress(indirect_src.value.value),
+                )
         elif dest.operation == LowLevelILOperation.LLIL_REG:
             if dest.value.type in [
                 RegisterValueType.ImportedAddressValue,
                 RegisterValueType.ConstantValue,
                 RegisterValueType.ConstantPointerValue,
             ]:
-                yield Characteristic("calls from"), AbsoluteVirtualAddress(dest.value.value)
+                yield (
+                    Characteristic("calls from"),
+                    AbsoluteVirtualAddress(dest.value.value),
+                )
 
 
 def extract_function_indirect_call_characteristic_features(
@@ -550,18 +643,26 @@ def extract_function_indirect_call_characteristic_features(
     ]:
         return
 
-    if llil.dest.operation in [LowLevelILOperation.LLIL_CONST, LowLevelILOperation.LLIL_CONST_PTR]:
+    if llil.dest.operation in [
+        LowLevelILOperation.LLIL_CONST,
+        LowLevelILOperation.LLIL_CONST_PTR,
+    ]:
         return
 
     if llil.dest.operation == LowLevelILOperation.LLIL_LOAD:
         src = llil.dest.src
-        if src.operation in [LowLevelILOperation.LLIL_CONST, LowLevelILOperation.LLIL_CONST_PTR]:
+        if src.operation in [
+            LowLevelILOperation.LLIL_CONST,
+            LowLevelILOperation.LLIL_CONST_PTR,
+        ]:
             return
 
     yield Characteristic("indirect call"), ih.address
 
 
-def extract_features(f: FunctionHandle, bbh: BBHandle, insn: InsnHandle) -> Iterator[tuple[Feature, Address]]:
+def extract_features(
+    f: FunctionHandle, bbh: BBHandle, insn: InsnHandle
+) -> Iterator[tuple[Feature, Address]]:
     """extract instruction features"""
     for inst_handler in INSTRUCTION_HANDLERS:
         for feature, ea in inst_handler(f, bbh, insn):